requirements.txt
anyio==3.6.1
asgiref==3.5.2
bleach==5.0.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-log==0.4.0
colorama==0.4.4
cryptography==37.0.2
deprecation==2.1.0
docutils==0.18.1
dotty-dict==1.3.0
fastapi==0.78.0
gitdb==4.0.9
GitPython==3.1.27
gotrue==0.5.0
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.21.3
idna==3.3
importlib-metadata==4.11.4
invoke==1.7.1
jeepney==0.8.0
keyring==23.6.0
packaging==21.3
pkginfo==1.8.3
postgrest-py==0.10.2
pycparser==2.21
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-gitlab==3.5.0
python-semantic-release==7.28.1
PyYAML==6.0
readme-renderer==35.0
realtime==0.0.4
requests==2.28.0
requests-toolbelt==0.9.1
rfc3986==1.5.0
SecretStorage==3.3.2
semver==2.13.0
setuptools-scm==6.4.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
starlette==0.19.1
storage3==0.3.4
supabase==0.5.7
tomli==2.0.1
tomlkit==0.10.2
tqdm==4.64.0
twine==3.8.0
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
watchgod==0.8.2
webencodings==0.5.1
websockets==9.1
zipp==3.8.0
